Removed useless acitivities
