+++ app/src/main/AndroidManifest.xml
@@ -14,19 +14,13 @@
         android:theme="@style/AppTheme">
 
         <activity
-            android:name=".presentation.splash.SplashScreenActivity"
+            android:name=".presentation.main.MainActivity"
             android:label="@string/app_name"
-            android:theme="@style/SplashTheme">
+            android:theme="@style/SplashScreenTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-
-        <activity
-            android:name=".presentation.main.MainActivity"
-            android:label="@string/app_name"
-            android:theme="@style/AppTheme.NoActionBar">
         </activity>
     </application>
 </manifest>
