+++ app/src/main/java/sasd97/java_blog/xyz/yandexweather/WeatherApp.java
@@ -2,8 +2,10 @@ package sasd97.java_blog.xyz.yandexweather;
 
 import android.app.Application;
 import android.content.Context;
+import android.os.StrictMode;
 import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatDelegate;
+import android.util.Log;
 
 import com.evernote.android.job.JobManager;
 
@@ -13,11 +15,9 @@ import sasd97.java_blog.xyz.yandexweather.background.UpdateWeatherJob;
 import sasd97.java_blog.xyz.yandexweather.di.AppComponent;
 import sasd97.java_blog.xyz.yandexweather.di.DaggerAppComponent;
 import sasd97.java_blog.xyz.yandexweather.di.MainComponent;
-import sasd97.java_blog.xyz.yandexweather.di.SplashScreenComponent;
 import sasd97.java_blog.xyz.yandexweather.di.modules.AppModule;
 import sasd97.java_blog.xyz.yandexweather.di.modules.MainModule;
 import sasd97.java_blog.xyz.yandexweather.di.modules.NavigationModule;
-import sasd97.java_blog.xyz.yandexweather.di.modules.SplashScreenModule;
 
 /**
  * Created by alexander on 07/07/2017.
@@ -25,13 +25,14 @@ import sasd97.java_blog.xyz.yandexweather.di.modules.SplashScreenModule;
 
 public class WeatherApp extends Application {
 
+    private static final String TAG = WeatherApp.class.getCanonicalName();
+
     static {
         AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_AUTO);
     }
 
     private AppComponent appComponent;
     private MainComponent mainComponent;
-    private SplashScreenComponent splashScreenComponent;
 
     @Inject JobManager jobManager;
 
@@ -44,11 +45,24 @@ public class WeatherApp extends Application {
         super.onCreate();
         appComponent = buildAppComponent();
         getAppComponent().inject(this);
+        onDebug();
         onInit();
     }
 
     private void onInit() {
         UpdateWeatherJob.scheduleJob();
+    }
+
+    public void onDebug() {
+        if (!BuildConfig.DEBUG) return;
+        Log.i(TAG, "Application is in debug mode");
+
+        StrictMode.setThreadPolicy(
+                new StrictMode.ThreadPolicy.Builder()
+                .detectNetwork()
+                .penaltyDeath()
+                .build()
+        );
     }
 
     public AppComponent getAppComponent() {
@@ -61,20 +75,10 @@ public class WeatherApp extends Application {
         return mainComponent;
     }
 
-    public SplashScreenComponent getSplashScreenComponent() {
-        if (splashScreenComponent == null) splashScreenComponent
-                = appComponent.plusSplashScreenComponent(new SplashScreenModule());
-        return splashScreenComponent;
-    }
-
     private AppComponent buildAppComponent() {
         return DaggerAppComponent.builder()
                 .appModule(new AppModule(this))
                 .navigationModule(new NavigationModule())
                 .build();
-    }
-
-    private void setupTasks() {
-
     }
 }
