+++ app/src/main/java/sasd97/java_blog/xyz/yandexweather/di/AppComponent.java
@@ -9,7 +9,6 @@ import sasd97.java_blog.xyz.yandexweather.di.modules.JobModule;
 import sasd97.java_blog.xyz.yandexweather.di.modules.MainModule;
 import sasd97.java_blog.xyz.yandexweather.di.modules.NavigationModule;
 import sasd97.java_blog.xyz.yandexweather.di.modules.NetModule;
-import sasd97.java_blog.xyz.yandexweather.di.modules.SplashScreenModule;
 
 /**
  * Created by alexander on 09/07/2017.
@@ -21,5 +20,4 @@ public interface AppComponent {
     void inject(WeatherApp app);
 
     MainComponent plusMainComponent(MainModule module);
-    SplashScreenComponent plusSplashScreenComponent(SplashScreenModule module);
 }
