--- app/src/main/java/sasd97/java_blog/xyz/yandexweather/di/scopes/SplashScreenScope.java
@@ -1,15 +1,0 @@
-package sasd97.java_blog.xyz.yandexweather.di.scopes;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-import javax.inject.Scope;
-
-/**
- * Created by alexander on 09/07/2017.
- */
-
-@Scope
-@Retention(RetentionPolicy.RUNTIME)
-public @interface SplashScreenScope {
-}
