--- app/src/main/java/sasd97/java_blog/xyz/yandexweather/presentation/splash/SplashScreenView.java
@@ -1,10 +1,0 @@
-package sasd97.java_blog.xyz.yandexweather.presentation.splash;
-
-import com.arellomobile.mvp.MvpView;
-
-/**
- * Created by alexander on 09/07/2017.
- */
-
-public interface SplashScreenView extends MvpView {
-}
