--- app/src/test/java/sasd97/java_blog/xyz/yandexweather/ExampleUnitTest.java
@@ -1,17 +1,0 @@
-package sasd97.java_blog.xyz.yandexweather;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-/**
- * ResponseWeather local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() throws Exception {
-        assertEquals(4, 2 + 2);
-    }
-}
